Рефакторинг.
Разделил основной файл на файл с хэндлерами и файл с утилитами.
